+++ src/components/features/languages/index.ts
@@ -1,0 +1,1 @@
+export { default as LanguagesView } from './LanguagesView';
